refactor admin settings routes
