+++ src/routes/admin/actions/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+	import ActionsForm from "$lib/components/admin/ActionsForm.svelte";
+</script>
+
+<ActionsForm />
